+++ nextjs/app/transifex/context-example/page.tsx
@@ -3,39 +3,19 @@
 import Link from 'next/link';
 import React from 'react';
 import { useLocale } from '@transifex/react';
-import { tx } from '@transifex/native';
 import { SocialMediaCard } from './SocialMediaCard';
+import { LanguageSwitcher } from '../shared/LanguageSwitcher';
 import '../styles.css';
 
 export default function ContextExamplePage() {
   const locale = useLocale();
   const currentLocale = String(locale);
 
-  const handleLanguageChange = (langCode: string) => {
-    tx.setCurrentLocale(langCode);
-  };
-
-  const languages = [
-    { code: 'en', name: 'English' },
-    { code: 'es', name: 'Español' },
-    { code: 'fr', name: 'Français' },
-  ];
-
   return (
     <div className="container">
       <div className="header">
         <h1 className="heading-1">Context & Comments</h1>
-        <div className="lang-switcher">
-          {languages.map((lang) => (
-            <button
-              key={lang.code}
-              onClick={() => handleLanguageChange(lang.code)}
-              className={`btn-lang ${currentLocale === lang.code ? 'active' : ''}`}
-            >
-              {lang.name}
-            </button>
-          ))}
-        </div>
+        <LanguageSwitcher />
       </div>
 
       <p className="debug-info">
